+++ src/lib/components/DashJsonGrid.react.js
@@ -140,6 +140,8 @@ DashJsonGrid.propTypes = {
       'atomMaterial',
       'evaDark',
       'moonLight',
+      'inherit',
+      'unset',
     ]),
     PropTypes.exact({
       /**